refactor: separate test-utils classes into individual files
- Move MigrateTestFixture, TestDatabase, MigrateTestChange to separate files
- Move utility functions (cleanText, getFixturePath, FIXTURES_PATH) to utils/ subdirectory
- Update index.ts to only export from separate files
- Maintain all existing functionality and imports
