+++ packages/core/test-utils/MigrateTestChange.ts
@@ -1,0 +1,8 @@
+export interface MigrateTestChange {
+  name: string;
+  dependencies?: string[];
+  timestamp?: string;
+  planner?: string;
+  email?: string;
+  comment?: string;
+}
